+++ src/cautiousrobot/__init__.py
@@ -1,6 +1,6 @@
 from cautiousrobot.download import download_images
 from cautiousrobot.buddy_check import BuddyCheck
-from cautiousrobot.utils import downsample_and_save_image
+from cautiousrobot.downsample_and_save import downsample_and_save_image
 
 
 # Create instance of the class
